requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
motor>=3.3.2
beanie>=1.25.0
pydantic>=2.6.2
python-dotenv>=1.0.1
typing-extensions>=4.9.0
starlette-admin>=0.15.1